+++ src/resources/resource_user.py
@@ -1,103 +1,13 @@
-import uuid as uuid_
-from abc import abstractmethod
-
 from flask_restful import marshal_with, reqparse
-from rubix_http.exceptions.exception import NotFoundException, BadDataException, ForbiddenException
+from rubix_http.exceptions.exception import NotFoundException, ForbiddenException, BadDataException
 from rubix_http.resource import RubixResource
 from werkzeug.security import check_password_hash
 
 from src.models.enum import StateType
 from src.models.user.model_user import UserModel
-from src.resources.rest_schema.schema_user import user_all_attributes, user_return_fields, user_all_fields_with_children
-from src.resources.utils import encode_jwt_token, encrypt_password, decode_jwt_token, get_access_token, \
-    parse_user_update
-
-
-class UserResourceList(RubixResource):
-    parser = reqparse.RequestParser()
-    for attr in user_all_attributes:
-        parser.add_argument(attr,
-                            type=user_all_attributes[attr]['type'],
-                            required=user_all_attributes[attr].get('required', False),
-                            help=user_all_attributes[attr].get('help', None),
-                            store_missing=False)
-
-    @classmethod
-    @marshal_with(user_all_fields_with_children)
-    def get(cls):
-        return UserModel.find_all()
-
-    @classmethod
-    @marshal_with(user_return_fields)
-    def post(cls):
-        args = cls.parser.parse_args()
-        uuid = str(uuid_.uuid4())
-        user = UserModel(uuid=uuid, **args)
-        user.password = encrypt_password(user.password)
-        user.save_to_db()
-        return user
-
-
-class UserResource(RubixResource):
-    @classmethod
-    @marshal_with(user_all_fields_with_children)
-    def get(cls, **kwargs):
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException('User does not exist')
-        return user
-
-    @classmethod
-    @marshal_with(user_return_fields)
-    def patch(cls, **kwargs):
-        args = parse_user_update()
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.update(**args)
-        return user
-
-    @classmethod
-    def delete(cls, **kwargs):
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.delete_from_db()
-        return '', 204
-
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        raise NotImplementedError
-
-
-class UserResourceByUUID(UserResource):
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        return UserModel.find_by_uuid(kwargs.get('uuid'))
-
-
-class UserResourceByUsername(UserResource):
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        return UserModel.find_by_username(kwargs.get('username'))
-
-
-class UserVerifyResource(RubixResource):
-
-    @classmethod
-    def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('username', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        user: UserModel = UserModel.find_by_username(args['username'])
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.state = StateType.VERIFIED
-        user.commit()
-        return {'message': 'User has been verified successfully'}
+from src.resources.rest_schema.schema_user import user_all_fields_with_children, user_all_fields
+from src.resources.utils import get_access_token, decode_jwt_token, parse_user_update, encrypt_password, \
+    encode_jwt_token
 
 
 class UserChangePasswordResource(RubixResource):
@@ -132,6 +42,41 @@ class UserLoginResource(RubixResource):
         if not check_password_hash(user.password, args['password']):
             raise BadDataException('username and password combination is incorrect')
         return encode_jwt_token(user.uuid, user.username)
+
+
+class UserResource(RubixResource):
+
+    @classmethod
+    @marshal_with(user_all_fields_with_children)
+    def get(cls):
+        access_token = get_access_token()
+        uuid = decode_jwt_token(access_token).get('sub', '')
+        user = UserModel.find_by_uuid(uuid)
+        if not user:
+            raise NotFoundException("User does not exist")
+        return user
+
+    @classmethod
+    @marshal_with(user_all_fields)
+    def patch(cls):
+        access_token = get_access_token()
+        uuid = decode_jwt_token(access_token).get('sub', '')
+        user = UserModel.find_by_uuid(uuid)
+        if not user:
+            raise NotFoundException("User does not exist")
+        args = parse_user_update()
+        user.update(**args)
+        return user
+
+    @classmethod
+    def delete(cls):
+        access_token = get_access_token()
+        uuid = decode_jwt_token(access_token).get('sub', '')
+        user = UserModel.find_by_uuid(uuid)
+        if user is None:
+            raise NotFoundException("User does not exist")
+        user.delete_from_db()
+        return '', 204
 
 
 class UserAuthenticateResource(RubixResource):
